Clear the static-analysis warnings
Each one is a real reading of the code, not noise, so each is fixed rather than
silenced:

- `ChunkData::light_state_equals` incremented its cursor inside the same
  condition it read it in, which is exactly the kind of thing that reads as a
  bug even when the order is defined.
- `NbtReader` widened a tag byte through `int8_t` implicitly; a tag byte is
  signed, so the sign extension is now written down.
- The size caps in `InflateLimits` and `kMaxCells` multiplied in `unsigned int`
  before widening; the first factor is now `size_t`, so the product is.
- `decode_symbol` built a table index from an `int` sum; each side is widened on
  its own, where both are known non-negative.
- `mc_palette` built load-time messages by chaining `+`, allocating a temporary
  per operator; a small `joined()` reserves once.

`crash_dump` moves its `getenv` to install time. That was not just a warning:
the flag is consulted from a fault path, and `getenv` takes a lock inside the C
runtime, so reading it there could have hung instead of reporting. Verified
armed end to end - the deliberate fault still writes its report, dump and
first-chance file.

498 tests pass, clean build.

+++ src/schematic/gzip_inflate.hpp
@@ -37,7 +37,7 @@ enum class WrapKind : uint8_t {
 // output from a tiny input, so the decompressed size is capped rather than
 // trusted: a schematic is a build, not a disk image.
 struct InflateLimits {
-    size_t max_output_bytes = 512u * 1024u * 1024u;
+    size_t max_output_bytes = static_cast<size_t>(512) * 1024u * 1024u;
 };
 
 // Inflates one member of the given container kind, appending to `out`.
